Move FileSubmission to subdir
