+++ indico/web/client/js/react/components/files/FileSubmission.jsx
