+++ indico/web/client/js/react/components/index.js
@@ -30,7 +30,7 @@ export {
 export {default as PopoverDropdownMenu} from './PopoverDropdownMenu';
 export {default as ResponsivePopup} from './ResponsivePopup';
 export {default as MathJax} from './MathJax';
-export {default as FileSubmission} from './FileSubmission';
+export {default as FileSubmission} from './files/FileSubmission';
 export {default as ClipboardButton} from './ClipboardButton';
 export {default as RequestConfirm} from './RequestConfirm';
 export {default as ReviewRating} from './ReviewRating';
